refactoring of eval_cps with readability as the goal

+++ src/heap.c
@@ -113,7 +113,7 @@ int heap_init_addr(cons_t *addr, unsigned int num_cells) {
   NIL = enc_sym(symrepr_nil);
   RECOVERED = enc_sym(DEF_REPR_RECOVERED);
 
-  if (((uint32_t)addr % 8) != 0) return 0;
+  if (((uintptr_t)addr % 8) != 0) return 0;
   
   heap_init_state(addr, num_cells, false);
 
@@ -127,7 +127,7 @@ int heap_init(unsigned int num_cells) {
 
   cons_t *heap = (cons_t *)malloc(num_cells * sizeof(cons_t));
 
-  if (((uint32_t)heap % 8) != 0) return 0;
+  if (((uintptr_t)heap % 8) != 0) return 0;
 
   if (!heap) return 0;
   heap_init_state(heap, num_cells, true);
